ai_engine/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
qdrant-client==1.14.2
google-genai==1.10.0
python-dotenv==1.1.0
